MNS-Studios/NewsBoarder/requirements.txt
customtkinter>=5.2.0
requests>=2.31.0
pyinstaller>=6.0.0
pillow>=10.0.0
